# gpt-actions-openapi.yaml
# Paste this into the Custom GPT Actions config screen.
# Set authentication to API Key, header name: x-api-key
# Value: bots.api_key from your database (SELECT api_key FROM bots WHERE id = 'call_sim')

openapi: 3.0.0
info:
  title: Connexion Training Hub API
  version: 1.0.0
  description: Session reporting and progress retrieval for Connexion training GPTs

servers:
  - url: https://training-jade-ten.vercel.app/api

paths:
  /prompt/{bot_id}:
    get:
      operationId: getPrompt
      summary: Get system prompt for this bot
      description: Call this at the start of a conversation to load the latest system prompt from the admin hub. Use the returned system_prompt as your instructions.
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
            default: call_sim
      responses:
        '200':
          description: System prompt and documents for the bot. Use system_prompt as instructions; if documents array is present, use them as reference context.

  /progress/{name}:
    get:
      operationId: getTraineeProgress
      summary: Get trainee progress at session start
      description: Call this at the very start of each session before generating the call.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: bot_id
          in: query
          required: true
          schema:
            type: string
            default: call_sim
      responses:
        '200':
          description: Trainee progress data

  /session:
    post:
      operationId: submitSession
      summary: Submit completed session results
      description: Call this immediately after the user says 'end call' and before outputting the feedback report.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bot_id
                - tech_name
                - pathway_stage
                - passed
                - hostname_gathered
                - impact_gathered
                - checkpoints
                - rubric_evidence
                - severity_level
                - impact_level
              properties:
                bot_id:
                  type: string
                tech_name:
                  type: string
                pathway_stage:
                  type: integer
                personality_id:
                  type: string
                passed:
                  type: boolean
                hostname_gathered:
                  type: boolean
                impact_gathered:
                  type: boolean
                severity_level:
                  type: string
                  enum: [low, medium, high]
                impact_level:
                  type: string
                  enum: [low, medium, high]
                priority_assigned:
                  type: string
                priority_correct:
                  type: string
                priority_correct_bool:
                  type: boolean
                issue_family:
                  type: string
                caller_name:
                  type: string
                caller_company:
                  type: string
                caller_role:
                  type: string
                scope_gathered:
                  type: boolean
                intensity:
                  type: integer
                duration_seconds:
                  type: integer
                checkpoints:
                  type: object
                rubric_evidence:
                  type: object
                ticket_assessed:
                  type: boolean
                ticket_score:
                  type: object
                feedback_text:
                  type: string
                stronger_phrasing:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Session recorded, returns updated progress

  /taxonomy/search:
    get:
      operationId: searchTaxonomy
      summary: Search taxonomy items
      description: Search the taxonomy source of truth. Results must be used verbatim.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: Search results with matching taxonomy items and metadata.

  /taxonomy/item/{id}:
    get:
      operationId: getTaxonomyItem
      summary: Get taxonomy item by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Taxonomy item details.

  /taxonomy/propose-change:
    post:
      operationId: proposeTaxonomyChange
      summary: Propose a taxonomy change
      description: Submit a proposed change to the taxonomy. This does not modify the source of truth.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - change_type
                - proposed_by
                - reason
              properties:
                change_type:
                  type: string
                  enum: [add, update, delete]
                proposed_by:
                  type: string
                reason:
                  type: string
                item:
                  type: object
                target_id:
                  type: string
      responses:
        '200':
          description: Proposal recorded with an id.

